Bug fixes Updated reports Updated examples
